refactor: consolidate three date-utils.js into single shared module
Merges extractDateString and generateDateRange from main/date-utils.js,
and the formatDateTime wrapper from src/utils/date-utils.js into the
canonical shared/date-utils.js. Updates all consumers and test files to
import from shared/date-utils directly, then removes the two now-redundant
copies.

Closes #440
