+++ shared/date-utils.js
@@ -3,12 +3,15 @@
  * CommonJS format so main/ can require() it directly;
  * esbuild resolves it for the renderer bundle.
  *
- * Only generic formatting lives here — domain-specific date logic
- * (extractDateString, generateDateRange) stays in main/date-utils.js.
+ * Consolidates all date utilities previously spread across:
+ *   - shared/date-utils.js (formatDateTime)
+ *   - main/date-utils.js   (extractDateString, generateDateRange)
+ *   - src/utils/date-utils.js (wrapper)
  */
 
 const DATE_LOCALE = 'fr-FR';
 const TIME_FORMAT = { hour: '2-digit', minute: '2-digit' };
+const DAY_LABEL_FORMAT = { day: '2-digit', month: '2-digit' };
 
 /**
  * Build a "date time" label from a date string and optional timestamp.
@@ -24,4 +27,30 @@ function formatDateTime(date, timestamp) {
   return `${date}${time ? ' ' + time : ''}`;
 }
 
-module.exports = { DATE_LOCALE, formatDateTime };
+/**
+ * Extract the YYYY-MM-DD part from an ISO date string.
+ * @param {string|null} iso - ISO date string
+ * @returns {string|null} "YYYY-MM-DD" or null if falsy
+ */
+function extractDateString(iso) {
+  return iso ? iso.slice(0, 10) : null;
+}
+
+/**
+ * Generate an array of { date, label } objects for the last N days.
+ * @param {number} [days=30] - number of days
+ * @returns {Array<{date: string, label: string}>}
+ */
+function generateDateRange(days = 30) {
+  const now = new Date();
+  return Array.from({ length: days }, (_, i) => {
+    const d = new Date(now);
+    d.setDate(d.getDate() - (days - 1 - i));
+    return {
+      date: d.toISOString().slice(0, 10),
+      label: d.toLocaleDateString(DATE_LOCALE, DAY_LABEL_FORMAT),
+    };
+  });
+}
+
+module.exports = { DATE_LOCALE, formatDateTime, extractDateString, generateDateRange };
