+++ tests/main/date-utils.test.js
@@ -1,6 +1,5 @@
 import { describe, it, expect } from 'vitest';
-const { extractDateString, generateDateRange } = require('../../main/date-utils');
-const { formatDateTime } = require('../../shared/date-utils');
+const { extractDateString, generateDateRange, formatDateTime } = require('../../shared/date-utils');
 
 describe('date-utils', () => {
   describe('extractDateString', () => {
